mse: adapter: v4l2: Add workaround for remove duplicate buffer
refs #112125

This commit add workaround for remove duplicate buffer. This W/A
needed if using GStreamer v4l2sink with show-preroll-frame property
set true. Under that condition, the same data is queued in
the first buffer and the second buffer.

This setting is NOT recommended for MSE that expects encoded data.
Please set false to the show-preroll-frame property of v4l2sink.

Signed-off-by: Mitsuhiro Kimura <mitsuhiro.kimura.kc@renesas.com>
Signed-off-by: Masaru Nagai <masaru.nagai.vx@renesas.com>