mse: core: Fix the last picture may be not sent when running as a video talker
refs #148279

This commit fixes issue that incrementing the done_buf_cnt too much when
running as a talker with using the timer_interval. This issue may be
affected when running under the following use case.

- running as a Audio or MPEG2TS talker.
or
- running as a Video talker with using the fps_numerator/fps_denominator
  parameters.

Typically it appears as a phenomenon that the last picture may be not
sent when running as a video talker with using fps parameters.

Signed-off-by: Takeshi Miura <takeshi.miura.hz@hitachi-solutions.com>
Signed-off-by: Masaru Nagai <masaru.nagai.vx@renesas.com>