rcar: mse: Add extra error messages to the main module component
refs #368085

After code analysis it was discovered that some of the functions could
be populated with the additional error messages. We will try to cover
all the places where an error code is ignored, some function just
returns the error without any message that might indicate some
malfunction an so on. The main purpose of this patch is to make future
debug of the possible errors in the driver easier to achieve.

Signed-off-by: Ivan Bilych <ivan.bilych@siemens.com>
Signed-off-by: Tu Trinh <tu.trinh.zy@renesas.com>